requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0